Updated build.gradle to use actual Kotlin version and moved some classes around
