+++ build.gradle
@@ -2,7 +2,7 @@ group 'ru.spbau.mit'
 version '1.0-SNAPSHOT'
 
 buildscript {
-    ext.kotlin_version = '1.1.51'
+    ext.kotlin_version = '1.2.0'
 
     repositories {
         mavenCentral()
